Update to FLIP v1.7 (#44)
- Fixed memory leaks in the C++/CUDA/Python versions of FLIP.
- HDR-FLIP: Fixed crashes that followed from the reference's median being 0.
Automatic start and stop exposures in HDR-FLIP are computed using statistics of the reference image. For the stop exposure, we use the median luminance of the reference. When that is 0, we used to get undefined behavior or a crash. Version 1.7 fixes this by clamping the median to have a minimum value of FLT_EPSILON.